Solve caching problems in offsets_within_radius (#751)
* Solve caching problems in offsets_within_radius

* apply new r also when checking cache

* change tests -> this would be breaking

* remove float conversion

* use int instead of float in grid_multi offset keys

* use int instead of float in grid_single offsets keys

* use Int instead of real in calculate_offsets

* simplify code in calculate_offsets

* Update CHANGELOG.md

+++ src/spaces/grid_single.jl
@@ -12,8 +12,8 @@ struct GridSpaceSingle{D,P} <: AbstractGridSpace{D,P}
     stored_ids::Array{Int,D}
     metric::Symbol
     offsets_at_radius::Dict{Int,Vector{NTuple{D,Int}}}
-    offsets_within_radius::Dict{Float64,Vector{NTuple{D,Int}}}
-    offsets_within_radius_no_0::Dict{Float64,Vector{NTuple{D,Int}}}
+    offsets_within_radius::Dict{Int,Vector{NTuple{D,Int}}}
+    offsets_within_radius_no_0::Dict{Int,Vector{NTuple{D,Int}}}
 end
 
 """
@@ -31,8 +31,8 @@ function GridSpaceSingle(d::NTuple{D,Int}; periodic = true, metric = :chebyshev)
     s = zeros(Int, d)
     return GridSpaceSingle{D,periodic}(s, metric,
         Dict{Int,Vector{NTuple{D,Int}}}(),
-        Dict{Float64,Vector{NTuple{D,Int}}}(),
-        Dict{Float64,Vector{NTuple{D,Int}}}(),
+        Dict{Int,Vector{NTuple{D,Int}}}(),
+        Dict{Int,Vector{NTuple{D,Int}}}(),
     )
 end
 # Implementation of space API
